deprecate bokeh.util.notebook
